image_processor/pyproject.toml
[project]
name = "image-processor-utility"
version = "0.1.0"
description = "A package for processing images."
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "exif>=1.6.1",
    "ipython>=8.33.0",
    "piexif>=1.1.3",
    "pillow>=11.1.0",
    "rich>=13.0.0",
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
image-processor = "image_processor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["image_processor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
